refactor(embed): move driver JS to driver.js resource file [skip ci]
